+++ R/poissons.operations.R
@@ -4,6 +4,7 @@
 #' @name poissons.operations
 #' @param ListeStations Dataframe contenant un colonne "Nom" avec le code de la station (RHJ)
 #' @param CodeOperation Affichage du CodeOperation - \code{FALSE} (par défault)
+#' @param codeinventaire Affichage du codeinventaire - \code{FALSE} (par défault)
 #' @param Sortie Forme du dataframe de sortie - \code{Simple} (par défault), \code{Propre} (format diffusable, avec stations) ou \code{Complet} (tous les champs)
 #' @keywords poissons
 #' @import DBI
@@ -20,13 +21,15 @@
 poissons.operations <- function(
   ListeStations = data.frame(Nom = character(0)),
   CodeOperation = FALSE,
-  Sortie = c("Simple","Propre","Complet")
+  codeinventaire = FALSE,
+  Sortie = c("Simple", "Propre", "Complet")
 )
 {
 
-  ## Évaluation des choix
+  #### Évaluation des choix ####
   Sortie <- match.arg(Sortie)
   
+  #### Collecte des données ####
   ## Ouverture de la BDD ##
   ## Connexion à la BDD
   dbP <- BDD.ouverture(Type = "Poissons")
@@ -47,101 +50,33 @@ poissons.operations <- function(
   Operations$DateDebut.x <- ymd(Operations$datedebut.x)
   Operations$DateDebut.x <- format(Operations$datedebut.x, "%Y-%m-%d")
   
-  ## Extraction des stations ##
-  # Test si le nom existe bien, sinon message d'erreur et arrêt de la fonction #
-  
+  #### Test de cohérence ####
   if(dim(Operations %>% filter(nom %in% ListeStations$Nom))[1] == 0 & dim(ListeStations)[1] != 0)
     warning("Attention : station(s) absente(s) des opérations de la base de données")
   
-  ## Simplification ##
-  # Travail sur une seule station
-  if(dim(ListeStations)[1] != 0 & CodeOperation == F & Sortie == "Simple"){
-    Operations <- 
-      Operations %>%
-    select(nom, datedebut.x) %>% 
-    filter(nom %in% ListeStations$Nom) %>% 
-    dplyr::rename(Station = nom, Date = datedebut.x) %>% 
-    arrange(Station, Date)}
+  #### Nettoyage & reformatage ####
+  renommage <- c("CodeOperation" = "codeoperation", 
+                 # "CodeInventaire" = "codeinventaire", 
+                 "Station" = "nom", 
+                 "Date" = "datedebut.x", 
+                 "X" = "xlambert", 
+                 "Y" = "ylambert", 
+                 "TypeCoord" = "typelambert"
+  )
   
-  if(dim(ListeStations)[1] != 0 & CodeOperation == F & Sortie == "Propre"){
-    Operations <- 
-      Operations %>%
-      filter(nom %in% ListeStations$Nom) %>% 
-      select(nom, limiteamont, datedebut.x, modeechantillonnage, xlambert, ylambert, typelambert) %>% 
-      dplyr::rename(Station = nom, Date = datedebut.x, X = xlambert, Y = ylambert, TypeCoord = typelambert) %>% 
-      arrange(Station, Date)}
+  Operations_v2 <-
+    Operations %>% 
+    {if(ListeStations %>% nrow() != 0) filter(., nom %in% ListeStations$Nom) else .} %>%
+    {if(Sortie == "Simple") select(., codeinventaire, codeoperation, nom, datedebut.x) else .} %>%
+    {if(Sortie == "Propre") select(., codeinventaire, codeoperation, nom, limiteamont, datedebut.x, modeechantillonnage, xlambert, ylambert, typelambert) else .} %>%
+    {if(CodeOperation == F) select(., -codeoperation) else .} %>%
+    {if(codeinventaire == F) select(., -codeinventaire) else .} %>%
+    rename(any_of(renommage)) %>%
+    arrange(Station, Date)
   
-  if(dim(ListeStations)[1] != 0 & CodeOperation == F & Sortie == "Complet"){
-    Operations <- 
-      Operations %>%
-      filter(nom %in% ListeStations$Nom) %>% 
-      dplyr::rename(CodeOperation = codeoperation, Station = nom, Date = datedebut.x, X = xlambert, Y = ylambert, TypeCoord = typelambert) %>% 
-      arrange(Station, Date)}  
-  
-  if(dim(ListeStations)[1] != 0 & CodeOperation == T & Sortie == "Simple"){
-    Operations <- 
-      Operations %>%
-      select(nom, datedebut.x, codeoperation) %>% 
-      filter(nom %in% ListeStations$Nom) %>% 
-      dplyr::rename(CodeOperation = codeoperation, Station = nom, Date = datedebut.x) %>% 
-      arrange(Station, Date)}
-  
-  if(dim(ListeStations)[1] != 0 & CodeOperation == T & Sortie == "Propre"){
-    Operations <- 
-      Operations %>%
-      filter(nom %in% ListeStations$Nom) %>% 
-      select(codeoperation, nom, limiteamont, datedebut.x, modeechantillonnage, xlambert, ylambert, typelambert) %>% 
-      dplyr::rename(CodeOperation = codeoperation, Station = nom, Date = datedebut.x, X = xlambert, Y = ylambert, TypeCoord = typelambert) %>% 
-      arrange(Station, Date)}
-  
-  if(dim(ListeStations)[1] != 0 & CodeOperation == T & Sortie == "Complet"){
-    Operations <- 
-      Operations %>%
-      filter(nom %in% ListeStations$Nom) %>% 
-      dplyr::rename(CodeOperation = codeoperation, Station = nom, Date = datedebut.x, X = xlambert, Y = ylambert, TypeCoord = typelambert) %>% 
-      arrange(Station, Date)}
-  
-  # Travail sur l'ensemble des stations
-  if(dim(ListeStations)[1] == 0 & CodeOperation == F & Sortie == "Simple"){
-    Operations <- 
-      Operations %>%
-      select(nom, datedebut.x) %>% 
-      dplyr::rename(Station = nom, Date = datedebut.x) %>% 
-      arrange(Station, Date)}
-  
-  if(dim(ListeStations)[1] == 0 & CodeOperation == F & Sortie == "Propre"){
-    Operations <- 
-      Operations %>%
-      select(nom, limiteamont, datedebut.x, modeechantillonnage, xlambert, ylambert, typelambert) %>% 
-      dplyr::rename(Station = nom, Date = datedebut.x, X = xlambert, Y = ylambert, TypeCoord = typelambert) %>% 
-      arrange(Station, Date)}
-  
-  if(dim(ListeStations)[1] == 0 & CodeOperation == F & Sortie == "Complet"){
-    Operations <- 
-      Operations %>%
-      dplyr::rename(CodeOperation = codeoperation, Station = nom, Date = datedebut.x, X = xlambert, Y = ylambert, TypeCoord = typelambert) %>% 
-      arrange(Station, Date)}
-  
-  if(dim(ListeStations)[1] == 0 & CodeOperation == T & Sortie == "Simple"){
-    Operations <- 
-      Operations %>%
-      select(nom, datedebut.x, codeoperation) %>% 
-      dplyr::rename(CodeOperation = codeoperation, Station = nom, Date = datedebut.x) %>% 
-      arrange(Station, Date)}
-  
-  if(dim(ListeStations)[1] == 0 & CodeOperation == T & Sortie == "Propre"){
-    Operations <- 
-      Operations %>%
-      select(codeoperation, nom, limiteamont, datedebut.x, modeechantillonnage, xlambert, ylambert, typelambert) %>% 
-      dplyr::rename(CodeOperation = codeoperation, Station = nom, Date = datedebut.x, X = xlambert, Y = ylambert, TypeCoord = typelambert) %>% 
-      arrange(Station, Date)}
-  
-  if(dim(ListeStations)[1] == 0 & CodeOperation == T & Sortie == "Complet"){
-    Operations <- 
-      Operations %>%
-      dplyr::rename(CodeOperation = codeoperation, Station = nom, Date = datedebut.x, X = xlambert, Y = ylambert, TypeCoord = typelambert) %>% 
-      arrange(Station, Date)}
-
-  return(as_tibble(Operations))
+  data_sortie <- Operations_v2 %>% as_tibble()
+    
+  #### Sortie ####
+  return(data_sortie)
   
 } # Fin de la fonction